@@ -1,33 +1,25 @@
-import { useState } from 'react';
-import Container from 'react-bootstrap/Container';
+import logo from './logo.svg';
+import './App.css';
 
-import OrderConfirmation from './pages/confirmation/OrderConfirmation';
-import OrderEntry from './pages/entry/OrderEntry';
-import OrderSummary from './pages/summary/OrderSummary';
-
-import { OrderDetailsProvider } from './contexts/OrderDetails';
-
-export default function App() {
-  // orderPhase needs to be 'inProgress', 'review' or 'completed'
-  const [orderPhase, setOrderPhase] = useState('inProgress');
-
-  let Component = OrderEntry; // default to order page
-  switch (orderPhase) {
-    case 'inProgress':
-      Component = OrderEntry;
-      break;
-    case 'review':
-      Component = OrderSummary;
-      break;
-    case 'completed':
-      Component = OrderConfirmation;
-      break;
-    default:
-  }
-
+function App() {
   return (
-    <OrderDetailsProvider>
-      <Container>{<Component setOrderPhase={setOrderPhase} />}</Container>
-    </OrderDetailsProvider>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
   );
 }
+
+export default App;
